Python 3.10 support
Python 3.10 dropped support for the loop parameter to asyncio.sleep()
and gather(), the default loop is used instead. I haven't audited
whether any of these code paths now need to use set_default_loop(), but
tests pass, so... LGTM?

Support for using the default loop was only added in 3.7, so this drops
support for Python 3.6.

https://bugs.python.org/issue42392

Fixes: #137